requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch
torchvision

streamlit
streamlit-cropper

matplotlib
numpy
pandas
pillow
